% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_describe_device}
\alias{iot1clickdevicesservice_describe_device}
\title{Given a device ID, returns a DescribeDeviceResponse object describing
the details of the device}
\usage{
iot1clickdevicesservice_describe_device(DeviceId)
}
\arguments{
\item{DeviceId}{[required] The unique identifier of the device.}
}
\description{
Given a device ID, returns a DescribeDeviceResponse object describing
the details of the device.
}
\section{Request syntax}{
\preformatted{iot1clickdevicesservice$describe_device(
  DeviceId = "string"
)
}
}

\keyword{internal}
