% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickdevicesservice_operations.R
\name{iot1clickdevicesservice_list_tags_for_resource}
\alias{iot1clickdevicesservice_list_tags_for_resource}
\title{List all tags on an AWS IoT 1-Click resource}
\usage{
iot1clickdevicesservice_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The AWS IoT 1-Click resource with tags to be listed. This value is an
Amazon Resource Name (ARN).}
}
\description{
List all tags on an AWS IoT 1-Click resource.
}
\section{Request syntax}{
\preformatted{iot1clickdevicesservice$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
